% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{all_palunhcr_graphic}
\alias{all_palunhcr_graphic}
\title{UNHCR preferred graphic colors to be use as a main set for graphics creation.}
\format{
An object of class \code{character} of length 7.
}
\usage{
all_palunhcr_graphic
}
\description{
\code{pal_unhcr_graphic} provides an additional UNHCR color palette
that collects the preferred UNHCR colors for graphics as a data frame containing 6 colors.
}
\details{
See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
pal_unhcr_graphic
dim(pal_unhcr_graphic)  # 1 6

# Access by position:
pal_unhcr_graphic[1]    # color blue (as df)
pal_unhcr_graphic[[1]]  # color value "#0072BC"

# Access by name:
pal_unhcr_graphic["blue"]   # color "blue5" (as df)
pal_unhcr_graphic[["blue"]] # color value "#0072BC"

# Plotting palette:
seecol(pal_unhcr_graphic)

}
\seealso{
\code{\link{pal_unhcr}} for the default UNHCR color palette;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{pal_blue}},
\code{\link{pal_darkblue}},
\code{\link{pal_green}},
\code{\link{pal_grey}},
\code{\link{pal_red}},
\code{\link{pal_unhcr_graphic}},
\code{\link{pal_unhcr}},
\code{\link{pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
