% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{blue}
\alias{blue}
\title{UNHCR color blue.}
\format{
An object of class \code{character} of length 1.
}
\usage{
blue
}
\description{
\code{blue} provides the preferred color of \code{\link{pal_blue}}
(as an atomic HEX character value) and is defined as
\code{\link{pal_blue}[[5]]}.
}
\details{
See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
blue  # HEX character "#0072BC" (as value)
all.equal(blue, pal_blue[[5]])  # TRUE (same HEX values)

seecol(blue)  # view color and details

}
\seealso{
\code{\link{pal_blue}} for the corresponding color palette;
\code{\link{pal_unhcr}} for the UNHCR default color palette with 5 main colors of \code{\link{pal_blue}};
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{darkblue}},
\code{\link{green}},
\code{\link{grey}},
\code{\link{red}},
\code{\link{yellow}}
}
\concept{preferred colors}
\keyword{datasets}
