% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{darkblue}
\alias{darkblue}
\title{UNHCR color darkblue.}
\format{
An object of class \code{character} of length 1.
}
\usage{
darkblue
}
\description{
\code{darkblue} provides the preferred color of \code{\link{pal_darkblue}}
(as an atomic HEX character value) and is defined as
\code{\link{pal_darkblue}[[6]]}.
}
\details{
See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
darkblue  # HEX character "#18375F" (as value)
all.equal(darkblue, pal_darkblue[[6]])  # TRUE (same HEX values)

seecol(darkblue)  # view color and details

}
\seealso{
\code{\link{pal_darkblue}} for the corresponding color palette;
\code{\link{pal_unhcr}} for the UNHCR default color palette with all 5 colors of \code{\link{pal_blue}};
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{blue}},
\code{\link{green}},
\code{\link{grey}},
\code{\link{red}},
\code{\link{yellow}}
}
\concept{preferred colors}
\keyword{datasets}
