% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{green}
\alias{green}
\title{UNHCR color green.}
\format{
An object of class \code{character} of length 1.
}
\usage{
green
}
\description{
\code{green} provides the preferred color of \code{\link{pal_green}}
(as an atomic HEX character value) and is defined as
\code{\link{pal_green}[[5]]}.
}
\details{
See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
green  # HEX character "#00B398" (as value)
all.equal(green, pal_green[[5]])  # TRUE (same HEX values)

seecol(green)  # view color and details

}
\seealso{
\code{\link{pal_green}} for the corresponding color palette;
\code{\link{pal_unhcr}} for the UNHCR default color palette with all 5 colors of \code{\link{pal_blue}};
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{blue}},
\code{\link{darkblue}},
\code{\link{grey}},
\code{\link{red}},
\code{\link{yellow}}
}
\concept{preferred colors}
\keyword{datasets}
