% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{grey}
\alias{grey}
\title{UNHCR color grey.}
\format{
An object of class \code{character} of length 1.
}
\usage{
grey
}
\description{
\code{grey} provides the preferred color of \code{\link{pal_grey}}
(as an atomic HEX character value) and is defined as
\code{\link{pal_grey}[[4]]}.
}
\details{
See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
grey  # HEX character "#666666" (as value)
all.equal(grey, pal_grey[[4]])  # TRUE (same HEX values)

seecol(grey)  # view color and details

}
\seealso{
\code{\link{pal_grey}} for the corresponding color palette;
\code{\link{pal_unhcr}} for the UNHCR default color palette with 5 main colors of \code{\link{pal_blue}};
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show and use color palettes.

Other preferred colors: 
\code{\link{blue}},
\code{\link{darkblue}},
\code{\link{green}},
\code{\link{red}},
\code{\link{yellow}}
}
\concept{preferred colors}
\keyword{datasets}
