% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{pal_blue}
\alias{pal_blue}
\title{UNHCR blue color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
pal_blue
}
\description{
\code{pal_blue} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of
\code{\link{Blue}}).
}
\details{
See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
pal_blue
dim(pal_blue)  # 1 6

# Preferred color:
pal_blue[5]    # preferred (named) color "blue5" (as df)
pal_blue[[5]]  # preferred color value "#0072BC"

# Access by position:
pal_blue[5]    # named color "blue5" (as df)
pal_blue[[5]]  # color value "#0072BC"

# Access by name:
pal_unhcr["blue5"]    # color "blue5" (as df)
pal_unhcr[["blue5"]]  # color value "#0072BC"

# Plotting palette:
seecol(pal_blue)

}
\seealso{
\code{\link{pal_unhcr}} for the UNHCR default color palette with 5 main colors of \code{\link{pal_blue}};
\code{\link{pal_darkblue}} for an alternative blue UNHCR color palette;
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{all_palunhcr_graphic}},
\code{\link{pal_darkblue}},
\code{\link{pal_green}},
\code{\link{pal_grey}},
\code{\link{pal_red}},
\code{\link{pal_unhcr_graphic}},
\code{\link{pal_unhcr}},
\code{\link{pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
