% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{pal_darkblue}
\alias{pal_darkblue}
\title{UNHCR darkblue color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
pal_darkblue
}
\description{
\code{pal_darkblue} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of
\code{\link{darkblue}}).
}
\details{
See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
pal_darkblue
dim(pal_darkblue)  # 1 6
pal_darkblue[6]    # preferred (named) color "darkblue6"
pal_darkblue[[6]]  # preferred color "darkblue6" OR "#18375F"

# Plotting palette:
seecol(pal_darkblue)

}
\seealso{
\code{\link{pal_unhcr}} for the UNHCR default color palette with all 5 colors of \code{\link{pal_blue}};
\code{\link{pal_blue}} for an alternative blue UNHCR color palette;
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{all_palunhcr_graphic}},
\code{\link{pal_blue}},
\code{\link{pal_green}},
\code{\link{pal_grey}},
\code{\link{pal_red}},
\code{\link{pal_unhcr_graphic}},
\code{\link{pal_unhcr}},
\code{\link{pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
