% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{pal_green}
\alias{pal_green}
\title{UNHCR green color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
pal_green
}
\description{
\code{pal_green} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of
\code{\link{green}}).
}
\details{
See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
pal_green
dim(pal_green)  # 1 6
pal_green[5]    # preferred (named) color "green5"
pal_green[[5]]  # preferred color "green5" OR "#00B398"

# Plotting palette:
seecol(pal_green)

}
\seealso{
\code{\link{pal_unhcr}} for the UNHCR default color palette with all 5 colors of \code{\link{pal_blue}};
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{all_palunhcr_graphic}},
\code{\link{pal_blue}},
\code{\link{pal_darkblue}},
\code{\link{pal_grey}},
\code{\link{pal_red}},
\code{\link{pal_unhcr_graphic}},
\code{\link{pal_unhcr}},
\code{\link{pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
