% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{pal_grey}
\alias{pal_grey}
\title{UNHCR grey color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
pal_grey
}
\description{
\code{pal_grey} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of
\code{\link{black}}).
}
\details{
See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
pal_grey
dim(pal_grey)  # 1 6
pal_grey[4]    # preferred (named) color "grey4"
pal_grey[[4]]  # preferred color "grey4" OR "#666666"

# Plotting palette:
seecol(pal_grey)

}
\seealso{
\code{\link{pal_unhcr}} for the UNHCR default color palette with 5 main colors of \code{\link{pal_blue}};
\code{\link{pal_darkblue}} for alternative to grey UNHCR color palettes;
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{all_palunhcr_graphic}},
\code{\link{pal_blue}},
\code{\link{pal_darkblue}},
\code{\link{pal_green}},
\code{\link{pal_red}},
\code{\link{pal_unhcr_graphic}},
\code{\link{pal_unhcr}},
\code{\link{pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
