% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{pal_unhcr}
\alias{pal_unhcr}
\title{UNHCR primary color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 11 columns.
}
\usage{
pal_unhcr
}
\description{
\code{pal_unhcr} provides the default UNHCR color palette
as a data frame containing 11 colors.
}
\details{
This is the primary (web/sRGB) scale.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
pal_unhcr
dim(pal_unhcr)  # 1 11

# Access by position:
pal_unhcr[1]    # 1nd named color "blue5" (as df)
pal_unhcr[[1]]  # 2nd color value "#0072BC"

# Access by name:
pal_unhcr["blue5"]    # color "blue5" (as df)
pal_unhcr[["blue5"]]  # color value "#0072BC"

# Plotting palette:
seecol(pal_unhcr)

}
\seealso{
\code{\link{pal_blue}} for the UNHCR default blue palette;
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{all_palunhcr_graphic}},
\code{\link{pal_blue}},
\code{\link{pal_darkblue}},
\code{\link{pal_green}},
\code{\link{pal_grey}},
\code{\link{pal_red}},
\code{\link{pal_unhcr_graphic}},
\code{\link{pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
