% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{pal_unhcr_graphic}
\alias{pal_unhcr_graphic}
\title{UNHCR main colors from the brand book in a color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
pal_unhcr_graphic
}
\description{
\code{pal_unhcr_primary} provides an additional UNHCR color palette
that collects the primary UNHCR color as a data frame containing 4 colors.
}
\details{
See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
pal_unhcr_primary
dim(pal_unhcr_primary)  # 1 4

# Access by position:
pal_unhcr_primary[1]    # color blue (as df)
pal_unhcr_primary[[1]]  # color value "#0072BC"

# Access by name:
pal_unhcr_primary["blue"]   # color "blue5" (as df)
pal_unhcr_primary[["blue"]] # color value "#0072BC"

# Plotting palette:
seecol(pal_unhcr_primary)

}
\seealso{
\code{\link{pal_unhcr}} for the default UNHCR color palette;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{all_palunhcr_graphic}},
\code{\link{pal_blue}},
\code{\link{pal_darkblue}},
\code{\link{pal_green}},
\code{\link{pal_grey}},
\code{\link{pal_red}},
\code{\link{pal_unhcr}},
\code{\link{pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
