% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{pal_yellow}
\alias{pal_yellow}
\title{UNHCR yellow color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
pal_yellow
}
\description{
\code{pal_yellow} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of
\code{\link{yellow}}).
}
\details{
See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
pal_yellow
dim(pal_yellow)  # 1 6
pal_yellow[5]    # preferred (named) color "yellow5"
pal_yellow[[5]]  # preferred color "yellow5" OR "#FAEB00"

# Plotting palette:
seecol(pal_yellow)

}
\seealso{
\code{\link{pal_unhcr}} for the UNHCR default color palette with all 5 colors of \code{\link{pal_seeblau}};
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{all_palunhcr_graphic}},
\code{\link{pal_blue}},
\code{\link{pal_darkblue}},
\code{\link{pal_green}},
\code{\link{pal_grey}},
\code{\link{pal_red}},
\code{\link{pal_unhcr_graphic}},
\code{\link{pal_unhcr}}
}
\concept{color palettes}
\keyword{datasets}
