% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{red}
\alias{red}
\title{UNHCR color red.}
\format{
An object of class \code{character} of length 1.
}
\usage{
red
}
\description{
\code{red} provides the preferred color of \code{\link{pal_red}}
(as an atomic HEX character value) and is defined as
\code{\link{pal_red}[[5]]}.
}
\details{
See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
red  # HEX character "#EF4A60" (as value)
all.equal(red, pal_red[[5]])  # TRUE (same HEX values)

seecol(red)  # view color and details

}
\seealso{
\code{\link{pal_red}} for the corresponding color palette;
\code{\link{pal_unhcr}} for the UNHCR default color palette with all 5 colors of \code{\link{pal_blue}};
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{blue}},
\code{\link{darkblue}},
\code{\link{green}},
\code{\link{grey}},
\code{\link{yellow}}
}
\concept{preferred colors}
\keyword{datasets}
