% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_ggplot_util.R
\name{scale_y_comma}
\alias{scale_y_comma}
\title{Helper functions to stylize axis labels: commas on Y}
\usage{
scale_y_comma(
  name = waiver(),
  breaks = waiver(),
  minor_breaks = waiver(),
  guide = waiver(),
  n.breaks = NULL,
  labels,
  limits = NULL,
  expand = c(0.01, 0),
  oob = censor,
  na.value = NA_real_,
  trans = "identity",
  position = "left",
  sec.axis = waiver(),
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = ",",
  decimal.mark = ".",
  trim = TRUE,
  ...
)
}
\arguments{
\item{accuracy, scale, prefix, suffix, big.mark, decimal.mark, trim}{See
[scales::comma_format()] or [scales::percent_format()]}

\item{...}{passed on to [scales::comma_format()] or [scales::percent_format()]}
}
\description{
Helper functions to stylize axis labels: commas on Y
}
