% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_palette.R
\docType{data}
\name{test_pal}
\alias{test_pal}
\title{Define discrete palette}
\format{
An object of class \code{character} of length 4.
}
\usage{
test_pal
}
\description{
Test palette

This is a test palette inspired by stop-light colors
}
\examples{
{
    seecol(test_pal)  # view color and details
}
}
\references{
https://en.wikipedia.org/wiki/Traffic_light
}
\keyword{datasets}
