% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_palette.R
\docType{data}
\name{test_pal_div}
\alias{test_pal_div}
\title{Define three colors for endpoints of diverging continuous palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
test_pal_div
}
\description{
(high, middle, low)
}
\examples{
{
seecol(test_pal_div)  # view color and details
}
}
\keyword{datasets}
