% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_unhcr.R
\name{theme_foundation}
\alias{theme_foundation}
\title{Basic UNHCR theme for ggplot2.}
\usage{
theme_foundation(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size (optional, numeric).
Default: \code{base_size = 11}.
Recommendations use 11 for print and 16 for web}

\item{base_family}{Base font family (optional, character).
Default: \code{base_family = "Lato"}.
Options include \code{"mono"}, \code{"sans"} (default), and "serif".}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot title family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{legend_title_family, legend_title_face, legend_title_size}{axis title font family, face and size}

\item{legend_title}{legend title if `TRUE` add legend title; default `FALSE`}

\item{plot_margin}{plot margin (specify with `ggplot2::margin()`)}

\item{grid_col, axis_col}{grid & axis colors}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis_text_size}{font size of axis text}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}
}
\description{
\code{theme_unhcr} provides a basic \bold{UNHCR} theme
to use in \bold{ggplot2} commands.
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_ipsum()

# seminal bar chart

update_geom_font_defaults()

count(mpg, class) \%>\%
  ggplot(aes(class, n)) +
  geom_col() +
  geom_text(aes(label=n), nudge_y=3) +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 bar chart example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_ipsum(grid="Y") +
  theme(axis.text.y=element_blank())
}

}
\concept{plot functions}
