% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_all_pal}
\alias{unhcr_all_pal}
\title{Look-up list of palettes}
\format{
An object of class \code{character} of length 3.
}
\usage{
unhcr_all_pal
}
\description{
Look-up list of palettes
}
\keyword{datasets}
