% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_black}
\alias{unhcr_black}
\title{UNHCR color black}
\format{
An object of class \code{character} of length 1.
}
\usage{
unhcr_black
}
\description{
\code{unhcr_black} provides the preferred color of black

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
unhcr_black  # "#000000"
}
}
\seealso{
\code{\link{unhcr_pal_grey}} for the corresponding color palette;
\code{\link{unhcr_pal}} for the UNHCR default color palette with 5 main colors of \code{\link{unhcr_pal_blue}};
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show and use color palettes.

Other preferred colors: 
\code{\link{unhcr_blue}},
\code{\link{unhcr_darkblue}},
\code{\link{unhcr_green}},
\code{\link{unhcr_grey}},
\code{\link{unhcr_red}},
\code{\link{unhcr_white}},
\code{\link{unhcr_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
