% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_blue}
\alias{unhcr_blue}
\title{UNHCR color blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
unhcr_blue
}
\description{
Preferred color blue (as HEX character value): ----
\code{unhcr_blue} provides the preferred color of \code{\link{unhcr_pal_blue}}
(as an atomic HEX character value) and is defined as
\code{\link{unhcr_pal_blue}[[5]]}.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
unhcr_blue  # HEX character "#0072BC" (as value)
all.equal(unhcr_blue, unhcr_pal_blue[[5]])  # TRUE (same HEX values)

seecol(unhcr_blue)  # view color and details
}
}
\seealso{
\code{\link{unhcr_pal_blue}} for the corresponding color palette;
\code{\link{unhcr_pal}} for the UNHCR default color palette with 5 main colors of \code{\link{unhcr_pal_blue}};
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{unhcr_black}},
\code{\link{unhcr_darkblue}},
\code{\link{unhcr_green}},
\code{\link{unhcr_grey}},
\code{\link{unhcr_red}},
\code{\link{unhcr_white}},
\code{\link{unhcr_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
