% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_unhcr.R
\name{unhcr_discrete}
\alias{unhcr_discrete}
\title{Ggplot scale color discrete and continuous palette based on the pal_unhcr_graphic}
\usage{
unhcr_discrete()
}
\description{
Ggplot scale color discrete and continuous palette based on the pal_unhcr_graphic
}
\examples{
library(scales)
scales::show_col(unhcr_graphic()(9))
}
