% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_green}
\alias{unhcr_green}
\title{Preferred color darkblue (as HEX character value)}
\format{
An object of class \code{character} of length 1.
}
\usage{
unhcr_green
}
\description{
UNHCR color green.

\code{green} provides the preferred color of \code{\link{unhcr_pal_green}}
(as an atomic HEX character value) and is defined as
\code{\link{unhcr_pal_green}[[5]]}.

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
    unhcr_green  # HEX character '#00B398' (as value)
    all.equal(unhcr_green, unhcr_pal_green[[5]])  # TRUE (same HEX values)
    
    seecol(unhcr_green)  # view color and details
}
}
\seealso{
\code{\link{unhcr_pal_green}} for the corresponding color palette;
\code{\link{unhcr_pal}} for the UNHCR default color palette with all 5 colors of \code{\link{unhcr_pal_blue}};
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{unhcr_black}},
\code{\link{unhcr_blue}},
\code{\link{unhcr_darkblue}},
\code{\link{unhcr_grey}},
\code{\link{unhcr_red}},
\code{\link{unhcr_white}},
\code{\link{unhcr_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
