% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{unhcr_paged}
\alias{unhcr_paged}
\title{Create a statistical report with UNHCR Brand}
\usage{
unhcr_paged(..., css = c("unhcr-fonts", "unhcr-page", "unhcr"), 
    template = pkg_resource("html", "unhcr_paged.html"), csl = pkg_resource("csl", 
        "unhcr.csl"), highlight = NULL, pandoc_args = NULL)
}
\arguments{
\item{..., css, template, csl, highlight, pandoc_args}{Arguments passed to \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is similar to \code{\link{html_paged}}.
}
