% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_pal}
\alias{unhcr_pal}
\title{unhcr default (web/sRGB) Primary color palette}
\format{
An object of class \code{data.frame} with 1 rows and 11 columns.
}
\usage{
unhcr_pal
}
\description{
UNHCR primary color palette. \code{unhcr_pal}
provides the default UNHCR color palette
as a data frame containing 11 colors.

This is the primary (web/sRGB) scale.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
unhcr_pal
dim(unhcr_pal)  # 1 11

# Access by position:
unhcr_pal[1]    # 1nd named color "blue5" (as df)
unhcr_pal[[1]]  # 2nd color value "#0072BC"

# Access by name:
unhcr_pal["blue5"]    # color "blue5" (as df)
unhcr_pal[["blue5"]]  # color value "#0072BC"

# Plotting palette:
seecol(unhcr_pal)
}

}
\seealso{
\code{\link{unhcr_pal_blue}} for the UNHCR default blue palette;
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{unhcr_pal_blue}},
\code{\link{unhcr_pal_darkblue}},
\code{\link{unhcr_pal_graphic}},
\code{\link{unhcr_pal_green}},
\code{\link{unhcr_pal_grey}},
\code{\link{unhcr_pal_primary}},
\code{\link{unhcr_pal_red}},
\code{\link{unhcr_pal_yellow}}
}
\concept{color palettes}
\keyword{datasets}
