% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_pal_blue}
\alias{unhcr_pal_blue}
\title{UNHCR blue color palette}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
unhcr_pal_blue
}
\description{
BLUE:  PANTONE 300 C/U / CMYK 99 50 0 0
RGB 0 114 188 / HEX #0072BC
\code{unhcr_pal_blue} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of \code{\link{Blue}}).

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
unhcr_pal_blue
dim(unhcr_pal_blue)  # 1 6

# Preferred color:
unhcr_pal_blue[5]    # preferred (named) color "blue5" (as df)
unhcr_pal_blue[[5]]  # preferred color value "#0072BC"

# Access by position:
unhcr_pal_blue[5]    # named color "blue5" (as df)
unhcr_pal_blue[[5]]  # color value "#0072BC"

# Access by name:
unhcr_pal["blue5"]    # color "blue5" (as df)
unhcr_pal[["blue5"]]  # color value "#0072BC"

# Plotting palette:
seecol(unhcr_pal_blue)
}
}
\seealso{
\code{\link{unhcr_pal}} for the UNHCR default color palette with 5 main colors of \code{\link{unhcr_pal_blue}};
\code{\link{unhcr_pal_darkblue}} for an alternative blue UNHCR color palette;
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{unhcr_pal_darkblue}},
\code{\link{unhcr_pal_graphic}},
\code{\link{unhcr_pal_green}},
\code{\link{unhcr_pal_grey}},
\code{\link{unhcr_pal_primary}},
\code{\link{unhcr_pal_red}},
\code{\link{unhcr_pal_yellow}},
\code{\link{unhcr_pal}}
}
\concept{color palettes}
\keyword{datasets}
