% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_pal_darkblue}
\alias{unhcr_pal_darkblue}
\title{UNHCR darkblue color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
unhcr_pal_darkblue
}
\description{
darkblue
 CMYK 99 83 36 27
 RGB 24 55 95
 HEX #18375F
\code{unhcr_pal_darkblue} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of
\code{\link{unhcr_darkblue}}).

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
unhcr_pal_darkblue
dim(unhcr_pal_darkblue)  # 1 6
unhcr_pal_darkblue[6]    # preferred (named) color "darkblue6"
unhcr_pal_darkblue[[6]]  # preferred color "darkblue6" OR "#18375F"

# Plotting palette:
seecol(unhcr_pal_darkblue)
}
}
\seealso{
\code{\link{unhcr_pal}} for the UNHCR default color palette with all 5 colors of \code{\link{unhcr_pal_blue}};
\code{\link{unhcr_pal_blue}} for an alternative blue UNHCR color palette;
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{unhcr_pal_blue}},
\code{\link{unhcr_pal_graphic}},
\code{\link{unhcr_pal_green}},
\code{\link{unhcr_pal_grey}},
\code{\link{unhcr_pal_primary}},
\code{\link{unhcr_pal_red}},
\code{\link{unhcr_pal_yellow}},
\code{\link{unhcr_pal}}
}
\concept{color palettes}
\keyword{datasets}
