% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_pal_graphic}
\alias{unhcr_pal_graphic}
\title{Scale of all XX preferred colors for graphic creation}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
unhcr_pal_graphic
}
\description{
UNHCR preferred graphic colors to be use as a main set for graphics creation.
\code{unhcr_pal_graphic} provides an additional UNHCR color palette
that collects the preferred UNHCR colors for graphics as a data frame containing 6 colors.

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
    unhcr_pal_graphic
    dim(unhcr_pal_graphic)
    
    # Access by position:
    unhcr_pal_graphic[1]  # color blue (as df)
    unhcr_pal_graphic[[1]]  # color value '#0072BC'
    
    # Access by name:
    unhcr_pal_graphic["Blue"]  # color 'blue5' (as df)
    unhcr_pal_graphic[["Blue"]]  # color value '#0072BC'
    
    # Plotting palette:
    seecol(unhcr_pal_graphic)
}

}
\seealso{
\code{\link{unhcr_pal}} for the default UNHCR color palette;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{unhcr_pal_blue}},
\code{\link{unhcr_pal_darkblue}},
\code{\link{unhcr_pal_green}},
\code{\link{unhcr_pal_grey}},
\code{\link{unhcr_pal_primary}},
\code{\link{unhcr_pal_red}},
\code{\link{unhcr_pal_yellow}},
\code{\link{unhcr_pal}}
}
\concept{color palettes}
\keyword{datasets}
