% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_pal_green}
\alias{unhcr_pal_green}
\title{UNHCR greem color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
unhcr_pal_green
}
\description{
UNHCR green color palette.
GREEN:  CMYK 90 0 52 0 / RGB 0 179 152 / HEX #00B398
\code{pal_green} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of
\code{\link{green}}).

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
unhcr_pal_green
dim(unhcr_pal_green)  # 1 6
unhcr_pal_green[5]    # preferred (named) color "green5"
unhcr_pal_green[[5]]  # preferred color "green5" OR "#00B398"

# Plotting palette:
seecol(unhcr_pal_green)
}
}
\seealso{
\code{\link{unhcr_pal}} for the UNHCR default color palette with all 5 colors of \code{\link{unhcr_pal_blue}};
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{unhcr_pal_blue}},
\code{\link{unhcr_pal_darkblue}},
\code{\link{unhcr_pal_graphic}},
\code{\link{unhcr_pal_grey}},
\code{\link{unhcr_pal_primary}},
\code{\link{unhcr_pal_red}},
\code{\link{unhcr_pal_yellow}},
\code{\link{unhcr_pal}}
}
\concept{color palettes}
\keyword{datasets}
