% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_pal_primary}
\alias{unhcr_pal_primary}
\title{Scale of all 4 preferred colors, Other combinations of UNHCR color palettes}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
unhcr_pal_primary
}
\description{
\code{unhcr_pal_primary} provides an additional UNHCR color palette
that collects the primary UNHCR color as a data frame containing 4 colors.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
unhcr_pal_primary
dim(unhcr_pal_primary)  # 1 4

# Access by position:
unhcr_pal_primary[1]    # color blue (as df)
unhcr_pal_primary[[1]]  # color value "#0072BC"

# Access by name:
unhcr_pal_primary["Blue"]   # color "blue5" (as df)
unhcr_pal_primary[["Blue"]] # color value "#0072BC"

# Plotting palette:
seecol(unhcr_pal_primary)
}
}
\seealso{
\code{\link{unhcr_pal}} for the default UNHCR color palette;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{unhcr_pal_blue}},
\code{\link{unhcr_pal_darkblue}},
\code{\link{unhcr_pal_graphic}},
\code{\link{unhcr_pal_green}},
\code{\link{unhcr_pal_grey}},
\code{\link{unhcr_pal_red}},
\code{\link{unhcr_pal_yellow}},
\code{\link{unhcr_pal}}
}
\concept{color palettes}
\keyword{datasets}
