% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_pal_yellow}
\alias{unhcr_pal_yellow}
\title{UNHCR yellow color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
unhcr_pal_yellow
}
\description{
The secondary accent colour may be used in some circumstances
for a graphic element.
YELLOW: PANTONE PROCESS YELLOW C/U
  CMYK 3 1 100 0
  RGB 250 235 0
  HEX #FAEB00
  \code{unhcr_pal_yellow} provides an additional UNHCR color palette
as a data frame containing 6 colors (shades of
\code{\link{unhcr_yellow}}).

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
    unhcr_pal_yellow
    dim(unhcr_pal_yellow)  # 1 6
    unhcr_pal_yellow[5]  # preferred (named) color 'yellow5'
    unhcr_pal_yellow[[5]]  # preferred color 'yellow5' OR '#FAEB00'
    
    # Plotting palette:
    seecol(unhcr_pal_yellow)
}
}
\seealso{
\code{\link{unhcr_pal}} for the UNHCR default color palette with all 5 colors of \code{\link{unhcr_yellow}};
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other color palettes: 
\code{\link{unhcr_pal_blue}},
\code{\link{unhcr_pal_darkblue}},
\code{\link{unhcr_pal_graphic}},
\code{\link{unhcr_pal_green}},
\code{\link{unhcr_pal_grey}},
\code{\link{unhcr_pal_primary}},
\code{\link{unhcr_pal_red}},
\code{\link{unhcr_pal}}
}
\concept{color palettes}
\keyword{datasets}
