% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_red}
\alias{unhcr_red}
\title{Preferred color red (as HEX character value)}
\format{
An object of class \code{character} of length 1.
}
\usage{
unhcr_red
}
\description{
UNHCR color red.

\code{red} provides the preferred color of \code{\link{unhcr_pal_red}}
(as an atomic HEX character value) and is defined as
\code{\link{unhcr_pal_red}[[5]]}.

See \url{https://media.unhcr.org/archive/Data-visualisation-colour-palette-2CZ7A2HOVFP9.html} for details.
}
\examples{
{
    unhcr_red  # HEX character '#EF4A60' (as value)
    all.equal(unhcr_red, unhcr_pal_red[[5]])  # TRUE (same HEX values)
    seecol(unhcr_red)  # view color and details
}
}
\seealso{
\code{\link{unhcr_pal_red}} for the corresponding color palette;
\code{\link{unhcr_pal}} for the UNHCR default color palette with all 5 colors of \code{\link{unhcr_pal_blue}};
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{unhcr_black}},
\code{\link{unhcr_blue}},
\code{\link{unhcr_darkblue}},
\code{\link{unhcr_green}},
\code{\link{unhcr_grey}},
\code{\link{unhcr_white}},
\code{\link{unhcr_yellow}}
}
\concept{preferred colors}
\keyword{datasets}
