% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_ggplot_theme.R
\name{unhcr_scale_fill_discrete}
\alias{unhcr_scale_fill_discrete}
\alias{unhcr_scale_colour_discrete}
\title{Discrete  fill scales based on the UNHCR graphic palette}
\usage{
unhcr_scale_fill_discrete(...)
}
\description{
See [unhcr_discrete()].
}
\author{
ipda | unhcr.org | 2020
}
