% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_templ_paged.R
\name{unhcr_templ_paged}
\alias{unhcr_templ_paged}
\title{Create a paged HTML document suitable for printing with UNHCR Brand}
\usage{
unhcr_templ_paged(...)
}
\arguments{
\item{...}{Arguments passed to
\code{pagedown::\link[pagedown]{html_paged}}.}
}
\value{
An R Markdown output format.
}
\description{
Create a paged HTML document suitable for printing with UNHCR Brand
}
\references{
\url{https://pagedown.rbind.io}
}
