% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_templ_ppt.R
\name{unhcr_templ_ppt}
\alias{unhcr_templ_ppt}
\title{UNHCR ppt template

Loads additional style and template file}
\usage{
unhcr_templ_ppt(...)
}
\arguments{
\item{...}{additional arguments provided to \code{html_document}}
}
\description{
UNHCR ppt template

Loads additional style and template file
}
\references{
https://raw.githubusercontent.com/dr-harper/example-rmd-templates/master/R/my_html_format.R
}
