% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_deprecated.R
\name{unhcr_theme}
\alias{unhcr_theme}
\title{UNHCR [ggplot2] theme following brand recommendations}
\usage{
unhcr_theme(...)
}
\arguments{
\item{...}{parameters to `unhcrthemes::theme_unhcr`}
}
\description{
\code{theme_unhcr} provides a basic \bold{UNHCR} theme
to use in \bold{ggplot2} commands.
}
