% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_color_def.R
\docType{data}
\name{unhcr_yellow}
\alias{unhcr_yellow}
\title{Preferred color yellow (as HEX character value)}
\format{
An object of class \code{character} of length 1.
}
\usage{
unhcr_yellow
}
\description{
UNHCR color yellow.

\code{yellow} provides the preferred color of \code{\link{unhcr_pal_yellow}}
(as an atomic HEX character value) and is defined as
\code{\link{unhcr_pal_yellow}[[5]]}.

See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
{
unhcr_yellow  # HEX character "#FAEB00" (as value)
all.equal(unhcr_yellow, unhcr_pal_yellow[[5]])  # TRUE (same HEX values)
seecol(unhcr_yellow)  # view color and details
}

}
\seealso{
\code{\link{unhcr_pal_yellow}} for the corresponding color palette;
\code{\link{unhcr_pal}} for the unhcr default color palette with all 5 colors of \code{\link{unhcr_pal_yellow}};
\code{\link{unhcr_pal_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{unhcr_black}},
\code{\link{unhcr_blue}},
\code{\link{unhcr_darkblue}},
\code{\link{unhcr_green}},
\code{\link{unhcr_grey}},
\code{\link{unhcr_red}},
\code{\link{unhcr_white}}
}
\concept{preferred colors}
\keyword{datasets}
