% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcRstyle-package.R
\docType{package}
\name{unhcRstyle-package}
\alias{unhcRstyle}
\alias{unhcRstyle-package}
\title{unhcRstyle: Use consistently UNHCR branding though preconfigured palette, ggplot2 themes and Rmd template}
\description{
A collection of template, palettes and ggplot2 themes, documents, and slides all based on UNHCR branding.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/unhcr-web/unhcRstyle}
  \item Report bugs at \url{https://github.com/unhcr-web/unhcRstyle/issues}
}

}
\author{
\strong{Maintainer}: Cedric Vidonne \email{vidonne@unhcr.org}

Authors:
\itemize{
  \item Edouard Legoupil \email{legoupil@unhcr.org} [contributor]
  \item Ahmadou Dicko \email{mail@ahmadoudicko.com} [contributor]
}

}
\keyword{internal}
