% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_unhcr.R
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_font_defaults(
  family = "Lato",
  face = "plain",
  size = 3.5,
  color = grey(0.15, 1)
)
}
\arguments{
\item{family, face, size, color}{font family name, face, size and color}
}
\description{
Updates [ggplot2::geom_label] and [ggplot2::geom_text] font defaults
}
