% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{using}
\alias{using}
\title{Load required packages for the notebook}
\usage{
using(...)
}
\arguments{
\item{...}{list of packages required}
}
\description{
Load required packages for the notebook
}
\examples{
\dontrun{
using("tidyverse", "gganimate", "gghighlight", "ggpubr", "dplyr", "tidyr")
}

}
