% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_def_unhcr.R
\docType{data}
\name{yellow}
\alias{yellow}
\title{UNHCR color yellow.}
\format{
An object of class \code{character} of length 1.
}
\usage{
yellow
}
\description{
\code{yellow} provides the preferred color of \code{\link{pal_yellow}}
(as an atomic HEX character value) and is defined as
\code{\link{pal_yellow}[[5]]}.
}
\details{
See \url{https://media.unhcr.org/archive/Brand-Book-for-external-partners-2CZ7A2P0CGAY.html} for details.
}
\examples{
yellow  # HEX character "#FAEB00" (as value)
all.equal(yellow, pal_yellow[[5]])  # TRUE (same HEX values)

seecol(yellow)  # view color and details

}
\seealso{
\code{\link{pal_yellow}} for the corresponding color palette;
\code{\link{pal_unhcr}} for the unhcr default color palette with all 5 colors of \code{\link{pal_seeblau}};
\code{\link{pal_unhcr_primary}} for a UNHCR color palette with primary colors;
\code{\link{seecol}} to show color palettes;
\code{\link{usecol}} to use color palettes.

Other preferred colors: 
\code{\link{blue}},
\code{\link{darkblue}},
\code{\link{green}},
\code{\link{grey}},
\code{\link{red}}
}
\concept{preferred colors}
\keyword{datasets}
