% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_ggplot_theme.R
\name{unhcr_discrete}
\alias{unhcr_discrete}
\title{Ggplot scale color discrete and continuous palette based on the unhcr_pal_graphic}
\usage{
unhcr_discrete()
}
\description{
Ggplot scale color discrete and continuous palette based on the unhcr_pal_graphic
}
\examples{
{
library(scales)
## to fix: the below...
#scales::show_col(unhcr_discrete(9))
}
}
\author{
ipda | unhcr.org | 2020
}
