% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_deprecated.R
\name{unhcr_templ_paged}
\alias{unhcr_templ_paged}
\title{Function for UNHCR paged report template}
\usage{
unhcr_templ_paged(...)
}
\arguments{
\item{...}{Arguments passed to `unhcrdown::paged_report`}
}
\value{
A pagedown report
}
\description{
Function for UNHCR paged report template
}
