% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_deprecated.R
\name{unhcr_templ_ppt}
\alias{unhcr_templ_ppt}
\title{Convert to UNHCR branded Powerpoint presentation}
\usage{
unhcr_templ_ppt(...)
}
\arguments{
\item{...}{parameters to pass to `unhcrdown::ppt_presentation`}
}
\description{
Format for converting from R Markdown to an UNHCR branded Powerpoint presentation
}
