% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_deprecated.R
\name{unhcr_templ_slide}
\alias{unhcr_templ_slide}
\title{Convert to UNHCR branded Xaringan presentation}
\usage{
unhcr_templ_slide(...)
}
\arguments{
\item{...}{parameters to pass to `unhcrdown::xaringan_presentation`}
}
\description{
Format for converting from R Markdown to an UNHCR branded Xaringan presentation
}
